refactor(biz/release): move programmable gateway release into separate file
