+++ src/dashboard/apigateway/apigateway/tests/controller/publisher/test_publish.py
@@ -67,9 +67,7 @@ class TestTriggerGatewayPublish:
         source = PublishSourceEnum.BACKEND_UPDATE
         ok, msg = _is_gateway_ok_for_releasing(fake_release, source)
         assert ok is False
-        assert (
-            f"The data structure of version 【{fake_release.resource_version.object_display}】 is incompatible" in msg
-        )
+        assert f"The data structure of version [{fake_release.resource_version.object_display}] is incompatible" in msg
 
     def test__save_release_history(self, fake_release):
         source = PublishSourceEnum.BACKEND_UPDATE
